Add files via upload
This project is a home automation system that integrates with sensors and leverages the power of IIOT (Industrial Internet of Things) technology and the Adafruit cloud to connect any smart door lock. The system is designed to provide a seamless and convenient experience for users, allowing them to remotely control and monitor their home from anywhere.